Added ability to show more reviews on home view
Added a button at the bottom of the home view which enables the user to
show more of the most recent reviews. On the initial get request, all
reviews are store in store.js, in chunks of three reviews, as well as a
global incrementer variable. each time the button is clicked, the
incrementer is incremented by one, and the next 3 reviews in the array
are loaded. If the incrementer reaches the length of the array, it
removes the button from the page so the user can no longer click it.
When a user signs in and out again, the incrementer is reset, the
reviews are called again overwriting the ones in store.js and the
actions can be performed all over again.